feat: Add iterative deepening
