+++ server/ai/ai_player.py
@@ -1,6 +1,5 @@
 import random
-import utils
-from ai import evaluator
+from ai import minimax
 import numpy as np
 import torch
 from network.q_learning import QNetwork, QNetworkAgent
@@ -21,22 +20,12 @@ class MinimaxPlayer:
         return self.find_best_move(game.board_state)
     
     def find_best_move(self, board_state):
-        start = time.time()
-        best_val = -evaluator.INFINITY
-        best_move = None
-        
-        for move in utils.get_valid_moves(board_state, self.player):
-            # Áp dụng nước đi lên temp_board
-            row, col = move
-            next_board = utils.make_move(board_state, row, col, self.player)
-            # Đánh giá trạng thái sau khi đi
-            depth = evaluator.get_search_depth(board_state)
-            value = evaluator.minimax(next_board, depth, self.player, False, -evaluator.INFINITY, evaluator.INFINITY)
-            if value >= best_val:
-                best_val = value
-                best_move = move
-        end = time.time()
-        print(f"Minimax evaluation time: {end - start:.4f} seconds")
+        # start = time.time()
+        max_depth = 4
+        time_limit = 5
+        best_move = minimax.id_minimax(board_state, self.player, max_depth, time_limit)
+        # end = time.time()
+        # print(f"Iterative deepening evaluation time: {end - start:.4f} seconds")
         return best_move
